Fix Determination of CircleCI Validity
CircleCI now sets the value of the system environment variable
`CI_PULL_REQUEST` to an empty value. This means that Elixir will
determine the variable to be present and return an empty string
for its value; not `nil`. This change appropriately detects that
pull requests are signified by non-empty strings.